requirements.txt
tensorflow==2.10.0
streamlit==1.12.0
matplotlib==3.5.0
scikit-learn==1.1.0
seaborn==0.11.0
pillow==9.0.0
requests==2.28.0
flask==2.2.0
werkzeug==2.2.0
kaggle==1.5.12